requirements.txt
python-telegram-bot==21.4
google-generativeai==0.8.3
requests
python-dotenv
Pillow
python-chess
